2023/final-autobahn-cyberscape/examplat/deploy/backend/src/requirements.txt
alembic==1.9.2
aniso8601==9.0.1
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
docopt==0.6.2
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.1
Flask-RESTful==0.3.9
flask-restx==1.0.5
Flask-SQLAlchemy==3.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.1
pydocs==0.2
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.46
urllib3==1.26.14
Werkzeug==2.2.2